module4/ProxyPool/requirements.txt
environs==7.3.1
Flask==1.0.3
attrs==19.1.0
retrying==1.3.3
aiohttp==3.6.2
requests==2.22.0
loguru==0.4.1
pyquery==1.4.1
redis==3.4.1
